requirements.txt
asgiref==3.9.1
boto3==1.40.16
botocore==1.40.16
cffi==1.17.1
cryptography==45.0.6
Django==5.2.5
dnspython==2.7.0
jmespath==1.0.1
pycparser==2.22
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
